Clean up API architecture - Consolidate to /api directory
MAJOR REFACTOR: Eliminate duplicate API files and consolidate architecture

✅ MOVED: Real implementations from backend/api/ to /api/
  - api/playlists.js (with full documentation and 946 lines of logic)
  - api/library.js (comprehensive library management)
  - api/friends.js (complete friends system)
  - api/search-users.js (full user search functionality)

✅ FIXED: Import paths to use ../backend/lib/prisma.js
✅ UPDATED: Dev server to import from /api instead of backend/api
✅ REMOVED: Duplicate wrapper files and old backend API files

BENEFITS:
- Single source of truth for API logic
- Maintains excellent documentation from original files
- Works for both local development and Vercel deployment
- Eliminates code duplication and maintenance overhead

This preserves all the beautiful documentation and comprehensive
functionality while creating a clean, unified API architecture.
